Average Hazard Ratio Estimation

Function to obtain the average hazard ratio.

Usage

ahr_estimation(x_subg, dummy_subg, est_coef, h0, gamma = 1)

Arguments

x_subg

(matrix)
the matrix with the subgroup covariates.

dummy_subg

(matrix)
the dummy matrix with the subgroup covariates.

est_coef

(matrix)
the estimated coefficients from the fitted model.

h0

(numeric)
the vector with the cumulative baseline hazard.

gamma

(scalar)
numeric value defining the weights to obtain the average hazard ratio. Default is 1 (in this case the average hazard ratio obtained can be interpreted as the odds of concordance).

Value

Average hazard ratio

Examples

ahr_estimation(
  elastic_net_surv$design1, elastic_net_surv$dummy1,
  elastic_net_surv$est_coef, elastic_net_surv$h0
)
#> [1] 0.6500483