Subgroup Treatment Effect Horseshoe

Function to obtain the estimated posterior distribution of the subgroup treatment effects considering a horseshoe fitted model.

Usage

trt_horseshoe(object, gamma = 1, l = NULL, m = 50)

Arguments

object

(horseshoe)
the horseshoe object.

gamma

(scalar)
numeric value defining the weights to obtain the average hazard ratio. Default is 1 (in this case the average hazard ratio obtained can be interpreted as the odds of concordance). Just needed when using survival data.

l

(scalar)
the maximum value of time that wants to be studied to obtain the average hazard ratio. Default is the maximum value of time when there was an event. Just needed when using survival data.

m

(scalar)
the value that defines the equally spaced time points where the survival curves are going to be studied. Default is 50. Just needed when using survival data.

Value

Approximated posterior distribution of the subgroup treatment effects.

Examples

trt_horseshoe(horseshoe_fit_surv, m = 1)