bonsaiforest: Shrinkage Based Forest Plots

Subgroup analyses are routinely performed in clinical trial analyses. From a methodological perspective, two key issues of subgroup analyses are multiplicity (even if only predefined subgroups are investigated) and the low sample sizes of subgroups which lead to highly variable estimates, see e.g. Yusuf et al (1991) doi:10.1001/jama.1991.03470010097038 . This package implements subgroup estimates based on Bayesian shrinkage priors, see Carvalho et al (2019) https://proceedings.mlr.press/v5/carvalho09a.html. In addition, estimates based on penalized likelihood inference are available, based on Simon et al (2011) doi:10.18637/jss.v039.i05 . The corresponding shrinkage based forest plots address the aforementioned issues and can complement standard forest plots in practical clinical trial analyses.

See also

Author

Maintainer: Isaac Gravestock isaac.gravestock@roche.com

Authors:

Mar Vazquez Rabunal mar.vazquez_rabunal@roche.com

Daniel Sabanés Bové daniel.sabanes_bove@roche.com

Marcel Wolbers marcel.wolbers@roche.com

Other contributors:

F. Hoffmann-La Roche AG [copyright holder, funder]